tests/wpt/web-platform-tests/encoding-detection/tr-windows-1254-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>tr windows-1254 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1254", 'Expected windows-1254');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/tr-windows-1254-late.sub.html");
</script>
